// src/src/app/common/enums/key-code.enum.ts
export enum KeyCode {
	Backquote = 'Backquote',
	Backspace = 'Backspace',
	Escape = 'Escape',
	Delete = 'Delete',
	BacksLash = 'BacksLash',
	Enter = 'Enter',
	ControlLeft = 'ControlLeft',
	ControlRight = 'ControlRight',
	ShiftLeft = 'ShiftLeft',
	ShiftRight = 'ShiftRight',
	AltLeft = 'AltLeft',
	AltRight = 'AltRight',
	CapsLock = 'CapsLock',
	ContextMenu = 'ContextMenu',
	Tab = 'Tab',
	MetaLeft = 'MetaLeft',
	MetaRight = 'MetaRight',
	Space = 'Space',
	Digit0 = 'Digit0',
	Digit1 = 'Digit1',
	Digit2 = 'Digit2',
	Digit3 = 'Digit3',
	Digit4 = 'Digit4',
	Digit5 = 'Digit5',
	Digit6 = 'Digit6',
	Digit7 = 'Digit7',
	Digit8 = 'Digit8',
	Digit9 = 'Digit9',
	Minus = 'Minus',
	Equal = 'Equal',
	BracketLeft = 'BracketLeft',
	BracketRight = 'BracketRight',
	Semicolon = 'Semicolon',
	Quote = 'Quote',
	Coma = 'Coma',
	Period = 'Period',
	Slash = 'Slash',
	KeyQ = 'KeyQ',
	KeyW = 'KeyW',
	KeyE = 'KeyE',
	KeyR = 'KeyR',
	KeyT = 'KeyT',
	KeyA = 'KeyA',
	KeyS = 'KeyS',
	KeyD = 'KeyD',
	KeyF = 'KeyF',
	KeyG = 'KeyG',
	KeyZ = 'KeyZ',
	KeyX = 'KeyX',
	KeyC = 'KeyC',
	KeyV = 'KeyV',
	KeyB = 'KeyB',
	KeyY = 'KeyY',
	KeyU = 'KeyU',
	KeyI = 'KeyI',
	KeyO = 'KeyO',
	KeyP = 'KeyP',
	KeyH = 'KeyH',
	KeyJ = 'KeyJ',
	KeyK = 'KeyK',
	KeyL = 'KeyL',
	KeyN = 'KeyN',
	KeyM = 'KeyM',
}