// src/src/app/common/enums/key-finger.enum.ts
export enum KeyFinger {
	BigLeft = 'BigLeft',
	IndexLeft = 'IndexLeft',
	MiddleLeft = 'MiddleLeft',
	RingLeft = 'RingLeft',
	PinkyLeft = 'PinkyLeft',
	BigRight = 'BigRight',
	IndexRight = 'IndexRight',
	MiddleRight = 'MiddleRight',
	RingRight = 'RingRight',
	PinkyRight = 'PinkyRight',
	Indifference = 'Indifference',
	Big = 'Big',
}